+++ amqpcpp/outbuffer.h
@@ -122,7 +122,7 @@ public:
      *  @param  string  char* to the string
      *  @param  size    size of string
      */
-    void add(const char *string, uint32_t size)
+    void add(const char *string, size_t size)
     {
         memcpy(_current, string, size);
         _current += size;
